Refactor into a new class session file generation from edit database
