+++ src/Portal/Portal/Helpers/ISessionFileGenerator.cs
@@ -1,0 +1,10 @@
+﻿using System.Threading.Tasks;
+using Portal.Models;
+
+namespace Portal.Helpers
+{
+    public interface ISessionFileGenerator
+    {
+        Task<SessionFile> PopulateSessionFile(int processId, string userFullName);
+    }
+}
